add audio helpers convertToMix/Scratch
